@@ -27,7 +27,7 @@
     "cookies-next": "^4.1.1",
     "date-fns": "^4.1.0",
     "mongoose": "^8.14.1",
-    "next": "14.1.3",
+    "next": "14.2.35",
     "next-intl": "^3.11.0",
     "next-themes": "^0.4.6",
     "react": "^18.2.0",
@@ -58,7 +58,7 @@
     "@typescript-eslint/parser": "^8.33.0",
     "autoprefixer": "^10.0.1",
     "eslint": "^8",
-    "eslint-config-next": "14.1.3",
+    "eslint-config-next": "14.2.35",
     "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-prettier": "^5.4.0",
