@@ -1,14 +1,8 @@
 import { NextResponse } from 'next/server';
 
 export async function GET() {
-  try {
-    return NextResponse.json({
-      status: 'healthy',
-      timestamp: new Date().toISOString(),
-      environment: process.env.NODE_ENV || 'unknown',
-      region: process.env.VERCEL_REGION || 'unknown',
-    });
-  } catch (error) {
-    return NextResponse.json({ status: 'error', error: 'Health check failed' }, { status: 500 });
-  }
+  return NextResponse.json({
+    status: 'healthy',
+    timestamp: new Date().toISOString(),
+  });
 }
